reorder scopes to match list in GH PAT page
Very minor! I just moved the `gist` scope up so that the required scopes were in the same order as when you're looking through them to create a new token (I was going a bit cross-eyed double checking which ones I needed and which ones I had selected already). Thanks.