Add `run_attempt` and `previous_attempt_url` to `github_actions_repository_workflow_run`
The workflow run represented by this table/the GitHub API endpoint is
aggregate of all the attempts of the run, with details that are a
combination of the first attempt and the last attempt of the run.  When
dealing with runs that have multiple attempts, it's useful to know how
many attempts there have been and where to find the previous attempt.